moves main.py into rufas
