+++ RUFAS/_main_.py
@@ -1,0 +1,5 @@
+"""Allows RuFaS to be run with `python -m RUFAS`."""
+
+from RUFAS.main import main
+
+main()
