+++ tests/test_main.py
@@ -6,7 +6,7 @@ from unittest.mock import MagicMock
 import pytest
 from pytest_mock import MockerFixture
 
-from main import CaseInsensitiveArgumentAction, main, parse_gnu_args
+from RUFAS.main import CaseInsensitiveArgumentAction, main, parse_gnu_args
 from RUFAS.output_manager import LogVerbosity
